requirements.txt
selenium==4.19.0
yt-dlp==2024.3.10
google-api-python-client==2.121.0
google-auth-oauthlib==1.2.0
Pillow==10.2.0
